Finished font rendering code in controller.rs.
The controller currently looks up the glyph for the given char and gets
the corresponding HMetrics to figure out the advance for the glyph.

The rect_for method in FontCache finds texture coordinates for the given
codepoint.

These combined produce vertex data which is sent to the renderer.

A convenience function for adding charsets were added to
res/font/mod.rs. This takes a set of common charsets - i.e. Lowercase,
Uppercase, Punctuation etc. - and generates a slice of chars to be used
with the cache_glyphs function in FontCache.