Added rect_for implementation for cache.
Also added a 'tex()' function for the renderer controller. This uses the
rect_for function to get texture coordinates and generates vertex data
for the given texture.

Added a tex_ix field to the Vertex struct in the renderer module to help
differentiate between different textures.

Added a tex_type field to the Vertex struct to help the renderer know
whether the the tex_ix means a font index or a texture index.