Added gamma correction to textures and fonts.
Texture2d was being used for texture and font caches. Unfortunately, the
raw data stored in image files was in sRGB colour space (non-linear),
meaning it appeared a lot brighter when drawn to the screen.

When glium's SrgbTexture2d is used, it enables OpenGL's gamma
correction, making the image appear at the intended brightness.